김태욱 : 엔진 업데이트(파티클)

+++ KTW_DX/GameEngineCoreShader/ParticleHeader.fx
@@ -26,7 +26,7 @@ struct ParticleInfoData
     float Speed;
     
     // 나 죽었다 살았다.
-    uint iActive;
+    float iActive;
 };
 
 cbuffer ParticleUpdateInfo : register(b7)
